% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateNearest.r
\name{annotateNearest}
\alias{annotateNearest}
\title{Annotate Each Provided Differentialy Expressed Region (DER) with Its
Closest Annotation(s)}
\usage{
annotateNearest(
  features,
  annotations,
  outputDirectory = NULL,
  annotationsType = NULL,
  orderBy = "start",
  select = NULL,
  verbose = TRUE
)
}
\arguments{
\item{features}{A \code{GenomicRanges} object augmented with the DER column, which
indicates if the region is differentially expressed, as returned by
DiffSegR::dea().}

\item{annotations}{A \code{String}. Path to the annotation file (in GFF or GTF
format).}

\item{outputDirectory}{A \code{String}. Path to the output directory.}

\item{annotationsType}{A vector of \code{String}. Keep only matching annotations.}

\item{orderBy}{A \code{String}. The name of the column on which returned
annotations are sorted.}

\item{select}{A vector of \code{String}. Keep only columns matching these
names in the returned \code{Data.frame}. By default all columns from
mcols(dds) and mcols(annotations) are kept.}

\item{verbose}{A \code{Boolean}. Should all the operations performed be displayed ?}
}
\value{
Annotated differentialy expressed regions as \code{Data.frame} object.
}
\description{
\code{annotateNearest()} relies on user-specified annotations in GFF or GTF
format to annotate the DERs found during the differential expression analysis.
Seven classes of labels translate the relative positions of the stranded DER
and its closest annotation(s):
\itemize{
\item antisense ;
\item upstream ;
\item downstream ;
\item inside ;
\item overlapping 3’ ;
\item overlapping 5’ ;
\item overlapping 5’, 3’.
}
For the non-stranded DERs, the two classes of labels are: overlapping and
non-overlapping.
}
