% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counting.r
\name{counting}
\alias{counting}
\title{Quantifying Region Expression}
\usage{
counting(
  data,
  features,
  featureCountsType = "fromBam",
  featureCountsOtherParams = list(),
  verbose = TRUE
)
}
\arguments{
\item{data}{The \code{List} object returned by \code{\link[=newExperiment]{newExperiment()}}.}

\item{features}{A \code{GRanges} object that contains the segment boundaries
(returned by \code{\link[=segmentationLFC]{segmentationLFC()}}).}

\item{featureCountsType}{A \code{String}. Select how to summarize counts:
\itemize{
\item fromBam      : from bam files using \link[Rsubread:featureCounts]{Rsubread::featureCounts} ;
\item fromCoverage : from coverage profiles.
}}

\item{featureCountsOtherParams}{A \code{List}. Other paramters passed on to
\code{\link[Rsubread:featureCounts]{Rsubread::featureCounts()}}.}

\item{verbose}{A \code{Boolean}. Should all the operations performed be
displayed ?}
}
\description{
In brief, \code{counting()} quantify expression of regions found by segmenting
the per-base Log2-FC of each locus.

In details \code{counting()} calls  \code{\link[Rsubread:featureCounts]{Rsubread::featureCounts()}} which assigns to
the regions the mapped reads from each replicate of both the reference and the
alternative biological condition. By default a read is allowed to be assigned
to more than one region if it is found to overlap with more than one region.
\code{\link[Rsubread:featureCounts]{Rsubread::featureCounts()}} ends by building a count matrix. \code{counting()} then
returns regions and associated count matrix as \code{SummarizedExperiment} object.
Alternatively, users may opt to quantify the expression of each region
directly from the coverage profiles.
}
