% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.r
\name{coverage}
\alias{coverage}
\title{Calculate Coverage Profiles}
\usage{
coverage(data, coverageType = "average", verbose = TRUE)
}
\arguments{
\item{data}{The \code{List} object returned by \code{\link[=newExperiment]{newExperiment()}}.}

\item{coverageType}{A \code{String}. Select how to compute the coverage profiles:
\itemize{
\item fivePrime  : coverage profiles compute on 5' ends of reads ;
\item threePrime : coverage profiles compute on 3' ends of reads ;
\item average    : coverageFactory coverage profile compute on
average of 5' & 3' ends of reads ;
\item fullLength : coverage profiles compute on full length reads.
}}

\item{verbose}{A \code{Logical}. Should all the operations performed be displayed ?}
}
\description{
\code{coverage()} calculates the coverage profiles for loci specified by the user,
using data from BAM files.
}
