% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoise.r
\name{denoise}
\alias{denoise}
\title{denoise}
\usage{
denoise(dds, data)
}
\arguments{
\item{dds}{A \code{DESeqDataSet} object augmented with modelStart and modelEnd
columns.}

\item{data}{The \code{List} object returned by \code{\link[=loadData]{loadData()}}.}
}
\value{
A \code{DESeqDataSet} augmented with log2FCPerBase, log2RefPerBase,
log2OtherPerBase columns.
}
\description{
\code{denoise()} estimates the normalized denoised coverages and log2-FC
per-base.
}
