% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffStats.r
\name{diffStats}
\alias{diffStats}
\title{Statistics on differentialy expressed regions.}
\usage{
diffStats(dds)
}
\arguments{
\item{dds}{The \code{DESeqDataSet} object returned by \code{\link[=dea]{dea()}}.}
}
\description{
\code{diffStats()} computes the percentage of rejected regions / bases.
}
