% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportResults.r
\name{exportResults}
\alias{exportResults}
\title{Export DiffSegR Results}
\usage{
exportResults(
  data,
  features,
  sizeFactors,
  outputDirectory,
  genome,
  annotations = NA,
  igvVersion = "8",
  select = c(),
  verbose = TRUE
)
}
\arguments{
\item{data}{The \code{List} object returned by \code{\link[=newExperiment]{newExperiment()}}.}

\item{features}{The \code{GRanges} part of the object returned by
\code{\link[=dea]{dea()}}.}

\item{sizeFactors}{A \code{numeric}vector of normalization factors for samples.`}

\item{outputDirectory}{A \code{String}. Path to the output directory. One directory
by locus will be create within outputDirectory.}

\item{genome}{A \code{String}. Path to the genome on which reads have
been aligned.}

\item{annotations}{A \code{String}. Path to the annotation file.}

\item{igvVersion}{A \code{String}. What IGV version should be used in the XML
file.}

\item{select}{A vector of \code{String}. Additional information that appear for
each region in IGV. only columns from features matching these names are
kept.}
}
\description{
For each locus, \code{exportResults()} saves DERs, not-DERs, segmentation,
coverage profiles and Log2-FC per-base in formats readable (bedGraph,GFF3)
by genome viewers like the Integrative Genome Viewer (IGV). For IGV, export
also creates an IGV sessiong in XML that allows loading all tracks in
one click. exportResults() can operate on each locus in parallel.
}
