% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newExperiment.r
\name{newExperiment}
\alias{newExperiment}
\title{Define a New Experiment}
\usage{
newExperiment(
  sampleInfo,
  referenceCondition,
  otherCondition,
  loci,
  nbThreads = 1,
  nbThreadsByLocus = 1
)
}
\arguments{
\item{sampleInfo}{A \code{String}. Path to the file with
sample information. The file includes the following columns:
\itemize{
\item sample         : identifier for the sample ;
\item condition      : identifier for the condition ;
\item replicate      : identifier for the replicate ;
\item bam            : path to the bam file ;
\item coverage       : path to the coverage file (in rds format);
\item isPairedEnd    : logical indicating if the library contain
paired-end reads or not (TRUE/FALSE) ;
\item strandSpecific : integer indicating if a strand-specific analysis
should be performed. ‘0’ (unstranded), ‘1’ (stranded) and ‘2’ (reversely
stranded). WARNING: All samples must be consistently stranded or non-stranded.
} Alternatively, one may provide a \code{Data.frame} that contains the same columns.}

\item{referenceCondition}{A \code{String}. The reference condition of
the RNA-Seq experiment.}

\item{otherCondition}{A \code{String}. The alternative condition that will be
compare to the reference one.}

\item{loci}{A \code{String}. Path to the file with loci information. The file
includes the following columns:
\itemize{
\item seqid      : chromosome identifier (match one of the chromosome
in the bam) for the target genomic region ;
\item chromStart : start position for the target genomic region ;
\item chromEnd   : end position for the target genomic region ;
\item locusID    : unique id for the target genomic region.
} Alternatively, one may provide a \code{Data.frame} that contains the same columns.}

\item{nbThreads}{An \code{Integer}. The overall number of threads used for the
analysis.}

\item{nbThreadsByLocus}{An \code{Integer}. The number of threads dedicated to each
locus. nbThreadsByLocus cannot be greater than nbThreads. The number of loci
analyzed in parallel is equal to the divisor of nbThreads by nbThreadsByLocus.}
}
\value{
A \code{List}. The list containing loaded data.
}
\description{
\code{newExperiment()} allows for loading various pieces of information related to RNA-Seq
samples and the loci that will be analyzed.
}
