% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postHocTest.r
\name{postHocTest}
\alias{postHocTest}
\title{Multiple Tests Correction Through Post-hoc Inference}
\usage{
postHocTest(SExp, predicate, alpha, tdpLowerBound, verbose = TRUE)
}
\arguments{
\item{SExp}{An object that inherits from the \code{SummarizedExperiment} class and
augmented with the results of the differential expression analysis.}

\item{predicate}{A predicate function that returns a single TRUE or FALSE if
the regions on which it is applied meets the conditions defined in the
predicate. The criteria used in the predicate have to be defined for each
regions in mcols(SExp).}

\item{alpha}{A \code{Double}. The significance level of the (post-hoc) confidence
bound on the type 1 error (See \link[sanssouci:posthocBySimes]{sanssouci::posthocBySimes}).}

\item{tdpLowerBound}{A value for the minimum true positive rate
on the returned set of rejected regions.}

\item{verbose}{A \code{Logical}. Should all the operations performed be displayed ?}
}
\value{
The \code{SExp} input object augmented with the metadata column
\code{DER}. \code{DER} is set to \code{TRUE} if the region is called differentially expressed.
}
\description{
Wrapper function around \code{sanssouci::posthocBySimes}
(Blanchard et al., 2020, doi:10.1214/19-AOS1847) for multiple tests correction
through post-hoc inference.
}
