% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateSD.R
\name{SD}
\alias{SD}
\title{Some robust SD estimators.}
\usage{
SD(y, w = NULL, method = "MAD")
}
\arguments{
\item{y}{A vector of \code{Double}. Observations.}

\item{w}{A vector of \code{Double}. Weights associated to the observations.}

\item{method}{A \code{String}. The type of estimator:
\itemize{
\item "MAD"
\item "wtd_MAD"
\item "HALL"
}}
}
\value{
SD estimation of y.
}
\description{
Robust SD estimators: MAD, weighted MAD, HALL.
}
