% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentationLFC.r
\name{segmentationLFC}
\alias{segmentationLFC}
\title{Segment the Per-base Log2-FC of All Loci}
\usage{
segmentationLFC(
  data,
  modelSelectionType = "yao",
  outputDirectory = NULL,
  alpha = 2,
  segmentNeighborhood = FALSE,
  Kmax = NULL,
  alphas = NULL,
  gridSearch = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{The \code{List} object returned by \code{\link[=newExperiment]{newExperiment()}}.}

\item{modelSelectionType}{A \code{String}. Select the penalty used by FPOP:
\itemize{
\item yao : Yao's penalty \code{alpha*sigma^2*log(n)}.
}}

\item{outputDirectory}{A \code{String}. Path to the output directory. On directory
by locus will be create within \code{outputDirectory}.}

\item{alpha}{A \code{Double}. Segmentation hyperparameter used in Yao's penalty:
\code{alpha*sigma^2*log(n)}. The number of changepoints returned by
\code{\link[fpopw:Fpop_w]{fpopw::Fpop_w()}} is a decreasing function of \code{alpha}.}

\item{segmentNeighborhood}{A \code{Logical}. Indicate the weighted pDPA algorithm
(Rigaill 2010 and 2015) has to be used to explore the space of segmentations.}

\item{Kmax}{An \code{Integer}. Segmentations with 1 to Kmax segments are recovered
using segment neighborhood.}

\item{alphas}{A vector of \code{Double}. A series of alphas used by the grid
search procedure.}

\item{gridSearch}{A \code{Logical}. Indicate if a grid search has to be used
to explore the space of segmentations.}

\item{verbose}{A \code{Logical}. Should all the operations performed be
displayed ?}
}
\value{
A \code{GRanges} object that contains the segment boundaries.
}
\description{
WARNING : \code{\link[=coverage]{coverage()}} must be run before \code{segmentationLFC()}.

In brief, \code{segmentationLFC()} identifies segments with homogeneous Log2-FC
along the loci of interest. segmentationLFC() can operate on each locus in
parallel.

In details, for each locus, \code{segmentationLFC()} calls \code{\link[fpopw:Fpop_w]{fpopw::Fpop_w()}} on the
Log2-FC profile (possibly stranded) to retreive the segments boundaries
(changepoints). The reference condition specified by the user is set to
the denominator of the Log2-FC. The hyperparameter \code{alpha} specified by
regions are returned as \code{GRanges} object.
}
