% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPaths.r
\name{setPaths}
\alias{setPaths}
\title{Update IGV Sessions}
\usage{
setPaths(directory, genome, annotations)
}
\arguments{
\item{directory}{A \code{String}. Path to the directory containing the xml files.}

\item{genome}{A \code{String}. Path to the genome on which reads have been aligned.}

\item{annotations}{A \code{String}. Path to the annotation file (in GFF or GTF
format).}
}
\description{
This function allows updating the path to the reference genome and annotations
for each IGV session (in XML format) within a user-specified directory. XML
files are searched for in a recursive manner.
}
