% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea.r
\name{dea}
\alias{dea}
\title{Differential Expression Analysis (DEA)}
\usage{
dea(
  SExp,
  design = ~condition,
  sizeFactors = NA,
  significanceLevel = 0.05,
  predicate = NULL,
  postHoc_significanceLevel = 0.05,
  postHoc_tdpLowerBound = 0.95,
  verbose = TRUE
)
}
\arguments{
\item{SExp}{The \code{SummarizedExperiment} object returned by
\code{\link[=counting]{counting()}}.}

\item{design}{A \code{formula} object or a \code{Matrix} object. Passed on to
\link[DESeq2:DESeqDataSet]{DESeq2::DESeqDataSet}.}

\item{sizeFactors}{A vector of \code{Double}. Sample-specific size factors.}

\item{significanceLevel}{A \code{Double}. The significance cutoff on q-values.}

\item{predicate}{A predicate function that returns a single TRUE or FALSE if
the region on which it is applied meets the conditions defined in the
predicate. The criteria used in the predicate have to be defined for each
regions in mcols(SExp).}

\item{postHoc_significanceLevel}{A \code{Double}. The significance level of the
test procedure (See \link[sanssouci:posthocBySimes]{sanssouci::posthocBySimes}).}

\item{postHoc_tdpLowerBound}{A \code{Double}. The minimum true positive
proportion on the returned set of rejected regions.}

\item{verbose}{A \code{Logical}. Should all the operations performed be displayed ?}
}
\value{
A\code{DESeqDataSet} object augmented with the metadata column \code{DER}.
\code{DER} is set to \code{TRUE} if the region is called differentially expressed.
}
\description{
\code{dea} takes as input the recently built \code{SummarizedExperiment} object and calls
\code{DESeq2} to test for each line, i.e. each region, the difference in expression
between the two compared biological conditions. The user can choose to control
all the resulting p-values or only a subset of interest. In the first case
a Benjamini–Hochberg procedure is used to control the False Discovery Rate
(FDR). In the second case, the user specifies selection criteria for the
regions, e.g. a threshold on the absolute Log2-FC per-regions. Then, a
post-hoc procedure is used to control the joint Family wise error Rate
(jFWER) on the selected subset of regions.
}
