% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postHocTest.r
\name{postHocTest}
\alias{postHocTest}
\title{postHocTest}
\usage{
postHocTest(
  SExp,
  predicate,
  alpha,
  tdpLowerBound,
  orderBy = "pvalue",
  verbose = FALSE,
  dichotomicSearch = FALSE
)
}
\arguments{
\item{SExp}{An object that inherits from the \code{SummarizedExperiment} class and
augmented with the results of the differential expression analysis.}

\item{predicate}{A predicate function that returns a single TRUE or FALSE if
the hypothesis on which it is applied meets the conditions defined in the
predicate. The criteria used in the predicate have to be defined for each
hypotheses in mcols(SExp).}

\item{alpha}{A \code{Double}. The significance level of the (post-hoc) confidence
bound on the type 1 error (see \link[sanssouci:posthocBySimes]{sanssouci::posthocBySimes}).}

\item{tdpLowerBound}{A value for the minimum true positive rate
on the returned set of rejected regions.}

\item{orderBy}{A \code{String}. The parameter on which regions are ordered before
the post-hoc procedure.}

\item{verbose}{A \code{Boolean}. Should all the operations performed be displayed ?}

\item{dichotomicSearch}{A \code{Boolean}. Use it to speed up the post-hoc procedure.}
}
\value{
The \code{SummarizedExperiment} object returned by
\code{\link[=segmentation]{segmentation()}} augmented with the metadata column
\code{rejectedHypotheses}. The rejected regions belong to the largest set of
selected regions with a true positive rate of at least \code{tdpLowerBound}
on the \code{alpha} level.
}
\description{
Wrapper function around \code{sanssouci::posthocBySimes}
(Blanchard et al., 2020, doi:10.1214/19-AOS1847) for multiple tests correction
through post-hoc inference.
}
